Version 1.15 (#47)
* Create powershell-analysis.yml

* Add charset=utf-8 to the Content-Type header. (#40)

* Add charset=utf-8 to Content-Type header.

* Specify Content-Type as Invoke-ZoomRestMethod instead of New-ZoomHeaders.
Specifying Charset in the `-Headers` parameter of `Invoke-RestMethod` does not seem to have any effect.
We need to specify the `-ContentType` parameter.

* Fix the issue that New-ZoomMeetingPoll not working (#42)

* Change the Questions parameter of New-ZoomMeetingPoll to an array of IDictionary
The previous [string] array format did not follow the Zoom API specification and did not work.

* Split New-ZoomMeetingPollQuestion into a separate file.

* Chang the type of the Questions parameter of Update-ZoomMeetingPoll to IDictionary.
To match with New-ZoomMeetingPoll

* Add tests for New-ZoomMeetingPoll and New-ZoomMeetingPollQuestion

* Update Get-ZoomAccount.ps1

* Modifying cmdlet to be in line with best practices.

* Update Get-ZoomAccount.ps1

* Removing unneeded code.

* Removing unneeded code.

Write-Output should not be used for warnings. Consider verbose usage or error handling in Invoke-ZoomRestMethod.

* Remove unused code.

* Add ZoomAllMeeting and ZoomClougRecordingReport (#43)

* Added more examples.
Removed example that had code that didn't work.
Renamed to Get-ZoomMeetings in order to be me more in line with other cmdlets.
Changed type of $to and $from to datetime. Added some code to convert these to strings before passing to API call.
Fixed some issues with parameter sets.
Removed unused parameter $pagenumber.

* Removed Get-ZoomAllMeetings. Now Get-ZoomMeetings.

* Updated help for Type parameter.

* Update Get-ZoomMeetings.ps1

* Update Get-ZoomMeetings.ps1

* Update Get-ZoomMeetings.ps1

* Added ConvertTo-LoginTypeCode.ps1. Implemented it in cmdlets.

* Update README.md

* Create publishModule.yml

* Update PSZoom.psd1

Co-authored-by: mkht <mkht@users.noreply.github.com>
Co-authored-by: Malpractis <31232305+Malpractis@users.noreply.github.com>