Change PSZoomToken scope to Script
Fix: Store API Key/Secret in module scope variable #54

With the Server-to-Server OAuth token changes, and the $PSZoomToken variable being stored in the Global scope, it is possible for scripts to run over eachother (especially if you're doing a long running job) and overwrite the token being used. This change is designed to address this issue.